lld-wrapper: remove crt*.o if it has not been generated yet
During compilation of glibc (detection of ELF output format) lld tries
to access crt0.o and fails, because it has not yet been installed into
staging at this point. It is added due to the clang wrapper and is not
required as part of compilation.

Add a check to the lld wrapper to remove crt*.o if it has not yet been
installed to staging yet.

Signed-off-by: Lukas Auer <lukas.auer@aisec.fraunhofer.de>